fix: replace Int with Double on getting fixed card discount

+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/ui/dashspend/DashSpendViewModel.kt
@@ -604,7 +604,7 @@ class DashSpendViewModel @Inject constructor(
         exploreDao.getMerchantById(merchantId)
     }
 
-    fun getGiftCardDiscount(denomination: Int): Double {
+    fun getGiftCardDiscount(denomination: Double): Double {
         val merchantId = giftCardMerchant.value?.giftCardProviders?.find { it.provider == selectedProvider?.name}?.sourceId
         return merchantId?.let {
             when (selectedProvider) {
